start to split up and refactor
