+++ afskmodem.py
@@ -99,246 +99,7 @@ def log(level: int, data: str):
         if(LOG_TO_CONSOLE):
             print(output)
 
-"""
-    DigitalModes defines the speeds at which AFSKmodem can operate, 
-    and the tones that they use.
-"""
-class DigitalModes:
-    # Audio Frequency-Shift Keying (300 baud)
-    def afsk300() -> str: 
-        return "afsk300"
 
-    # Audio Frequency-Shift Keying (600 baud)
-    def afsk600() -> str: 
-        return "afsk600"
-    
-    # Audio Frequency-Shift Keying (1200 baud)
-    def afsk1200() -> str: 
-        return "afsk1200"
-
-    # Audio Frequency-Shift Keying (2400 baud)
-    def afsk2400() -> str: 
-        return "afsk2400"
-
-    # Audio Frequency-Shift Keying (6000 baud)
-    def afsk6000() -> str: 
-        return "afsk6000"
-
-    # Default (AFSK1200)
-    def default() -> str: 
-        return "afsk1200"
-    
-    # Returns the unit time in samples of the given mode
-    def getUnitTime(digital_mode: str) -> int:
-        if(digital_mode == "afsk300"):
-            return int(SAMPLE_RATE / 300)
-        elif(digital_mode == "afsk600"):
-            return int(SAMPLE_RATE / 600)
-        elif(digital_mode == "afsk1200"):
-            return int(SAMPLE_RATE / 1200)
-        elif(digital_mode == "afsk2400"):
-            return int(SAMPLE_RATE / 2400)
-        elif(digital_mode == "afsk6000"):
-            return int(SAMPLE_RATE / 6000)
-        else: # default
-            return int(SAMPLE_RATE / 1200)
-
-    # Returns the number of training sequence oscillations 
-    # for a specified time
-    def getTrainingCycles(sequence_time: int, digital_mode: str) -> int:
-        if(digital_mode == "afsk300"):
-            return int(300 * sequence_time / 2)
-        elif(digital_mode == "afsk600"):
-            return int(600 * sequence_time / 2)
-        elif(digital_mode == "afsk1200"):
-            return int(1200 * sequence_time / 2)
-        elif(digital_mode == "afsk2400"):
-            return int(2400 * sequence_time / 2)
-        elif(digital_mode == "afsk6000"):
-            return int(6000 * sequence_time / 2)
-        else: # default
-            return int(1200 * sequence_time / 2)
-    
-    # Get the frequency of the space tone for a given mode
-    def getSpaceTone(digital_mode: str) -> int:
-        if(digital_mode == "afsk300"):
-            return 300
-        elif(digital_mode == "afsk600"):
-            return 600
-        elif(digital_mode == "afsk1200"):
-            return 1200
-        elif(digital_mode == "afsk2400"):
-            return 2400
-        elif(digital_mode == "afsk6000"):
-            return 6000
-        else: # default
-            return 1200
-
-    # Get the frequency of the mark tone for a given mode
-    def getMarkTone(digital_mode: str) -> int:
-        if(digital_mode == "afsk300"):
-            return 600
-        elif(digital_mode == "afsk600"):
-            return 1200
-        elif(digital_mode == "afsk1200"):
-            return 2400
-        elif(digital_mode == "afsk2400"):
-            return 4800
-        elif(digital_mode == "afsk6000"):
-            return 12000
-        else: # default
-            return 2400
-
-"""
-    IdealWaves provides functions to load waveforms from files and return them 
-    as lists of amplitudes, stored as either bytes or ints.
-"""
-class IdealWaves:
-    def __init__(self, digital_mode = DigitalModes.default()):
-        self.digital_mode = digital_mode
-
-    # Load waveform data from a file 
-    # and convert it to a list of amplitudes as ints
-    def __loadWavFile(self, filename: str) -> list[int]:
-        with wave.open(filename, "r") as f:
-            nFrames = f.getnframes()
-            expFrames = []
-            for i in range(0, nFrames):
-                sFrame = f.readframes(1)
-                expFrames.append(struct.unpack("<h", sFrame)[0])
-            return expFrames
-
-    # Load waveform data from a file 
-    # and convert it to a list of amplitudes as bytes
-    def __loadRawWavFile(self, filename: str) -> bytes:
-        with wave.open(filename, "r") as f:
-            nFrames = f.getnframes()
-            return f.readframes(nFrames)
-    
-    # Silence (20ms) to pad output with for TX
-    def getTxSilence(self) -> bytes: 
-        return self.__loadRawWavFile(WAVEFORMS_DIR + "_.wav")
-    
-    # Space tone as bytes for transmitting
-    def getTxSpace(self) -> bytes: 
-        return self.__loadRawWavFile(WAVEFORMS_DIR + 
-                                     self.digital_mode + "/0.wav")
-    
-    # Mark tone as bytes for transmitting
-    def GetTxMark(self) -> bytes: 
-        return self.__loadRawWavFile(WAVEFORMS_DIR + 
-                                     self.digital_mode + "/1.wav")
-    
-    # Space tone as int array for receiving
-    def getRxSpace(self) -> list[int]: 
-        return self.__loadWavFile(WAVEFORMS_DIR + 
-                                  self.digital_mode + "/0.wav")
-    
-    # Mark tone as int array for receiving
-    def getRxMark(self) -> list[int]: 
-        return self.__loadWavFile(WAVEFORMS_DIR + 
-                                  self.digital_mode + "/1.wav")
-    
-    # Single training sequence cycle for RX clock recovery
-    def getRxTraining(self) -> list[int]: 
-        return self.getRxMark() + self.getRxSpace()
-
-"""
-    Hamming provides functions related to error correction functionality and
-    bit-error tracking.
-"""
-class Hamming:
-    # Each instance of Hamming keeps track of the errors it corrects. 
-    # An instance of Hamming is created for each DigitalTransmitter 
-    # or DigitalReceiver instance.
-    def __init__(self): 
-        self.r = 4
-        self.error_count = 0
-
-    # Reset error count to 0
-    def resetErrorCount(self):
-        self.error_count = 0
-    
-    # Get error count
-    def getErrorCount(self) -> int:
-        return self.error_count
-    
-    # Pad the positions of parity bits with 0s
-    def __padParityBits(self, data: str) -> str:
-        j = 0
-        k = 1
-        m = len(data)
-        p = ''
-        for i in range(1, m + self.r+1):
-            if(i == 2**j):
-                p += '0'
-                j += 1
-            else:
-                p += data[-1 * k]
-                k += 1
-        return p[::-1]
-
-    # Set the parity bits to their correct values
-    def __setParityBits(self, data: str) -> str:
-        n = len(data)
-        for i in range(self.r):
-            p = 0
-            for j in range(1, n + 1):
-                if(j & (2**i) == (2**i)):
-                    p = p ^ int(data[-1 * j])
-            data = data[:n-(2**i)] + str(p) + data[n-(2**i)+1:]
-        return data
-
-    # Find an error (if it exists)
-    def __getErrorIndex(self, data: str) -> int:
-        n = len(data)
-        p = 0
-        for i in range(self.r):
-            k = 0
-            for j in range(1, n + 1):
-                if(j & (2**i) == (2**i)):
-                    k = k ^ int(data[-1 * j])
-            p += k*(10**i)
-        return n - int(str(p), 2)
-
-    # Trim the parity bits off a corrected chunk to get the data only
-    def __trimParityBits(self, data: str) -> str:
-        data = data[::-1]
-        p = ""
-        j = 0
-        for i in range(len(data)):
-            if(i + 1 != 2 ** j):
-                p += data[i]
-            else:
-                j += 1
-        return p[::-1]
-
-    # Correct a single error in a section of data.
-    def __correctErrors(self, data: str) -> str:
-        error_pos = self.__getErrorIndex(data)
-        if(error_pos == len(data)):
-            return data
-        else:
-            self.error_count += 1
-            data_list = list(data)
-            if(data_list[error_pos] == "0"):
-                data_list[error_pos] = "1"
-            else:
-                data_list[error_pos] = "0"
-            data = "".join(data_list)
-        return data
-
-    # Encodes a section of data.
-    def encode(self, data: str) -> str:
-        padded_data = self.__padParityBits(data)
-        parity_data = self.__setParityBits(padded_data)
-        return(parity_data)
-
-    # Corrects and decodes a section of data.
-    def decode(self, data: str) -> str:
-        corrected_data = self.__correctErrors(data)
-        output_data = self.__trimParityBits(corrected_data)
-        return(output_data)
 
 """
     DigitalReceiver is a user-defined receiver class that provides functionality 
@@ -520,20 +281,6 @@ class DigitalReceiver:
             i += 8
         return bytes(int_data)
     
-    # Run error correction and remove all parity bits
-    def __correctErrors(self, data: str) -> str:
-        data_bytes = []
-        decoded_bytes = []
-        data_iter = 0
-        self.ecc.resetErrorCount()
-        while(data_iter < len(data) - 11):
-            data_bytes.append(data[data_iter:data_iter+12])
-            data_iter += 12
-        for i in data_bytes:
-            decoded_bytes.append(self.ecc.decode(i))
-        output = "".join(decoded_bytes)
-        return output, self.ecc.getErrorCount()
-    
     # Receives data and returns it (or fails)
     def rx(self, timeout = -1):
         log(0, "Receiver - listening...")
@@ -599,17 +346,7 @@ class DigitalTransmitter:
         return output
 
     # Generate ECC and insert into data.
-    def __generateECC(self, data: str) -> str:
-        data_bytes = []
-        encoded_bytes = []
-        dataIter = 0
-        while(dataIter < len(data) - 7):
-            data_bytes.append(data[dataIter:dataIter+8])
-            dataIter += 8
-        for dataByte in data_bytes:
-            encoded_bytes.append(self.ecc.encode(dataByte))
-        output = "".join(encoded_bytes)
-        return(output)
+
 
     # Play a waveform
     def __playWaveform(self, data: bytes):
